generator: Refactor to make github API calls only when needed
Currently, everytime the verify job is run, API calls are made
to github to fetch kubernetes release information since verify
calls generate first.

The issue with this is that everytime this runs in the CI, eventhough
we only generate annual reports when ANNUAL_REPORT env is set, the
API calls are made and if they are rate limited (since we don't use
a token here), the job will fail without a legitamate reason.

Signed-off-by: Madhav Jivrajani <madhav.jiv@gmail.com>