Add Krkn slack channel
This commit adds a slack channel for Krkn project - https://github.com/redhat-chaos/krkn,
a chaos testing tool for Kubernetes focusing on resilience and performance validation
under failure conditions.

This channel will act a discussion place for the Kubernetes community interested
in the Krkn tooling and methodology as well Chaos and Performance Engineering.