chore: add openstack-magnum channel
This is a request to add #openstack-magnum channel in the Kubernetes Slack.

Magnum is an OpenStack project which offers APIs and drivers
for deploying and managing Kubernetes as first class resources in OpenStack.

We'd like to have space to discuss with Kubernetes Cloud Provider for
OpenStack, CAPI, CAPO and other Kubernetes deployment related projects.