sig-cluster-lifecycle: move logical-cluster to a subproject repo
As the logical-cluster repository README indicates, this
project was primary created for serving in CAPI usage. Later
it can become used by other subprojects like kOps.

However, having it as a subproject creates extra work for us that
is not necessary - e.g. subproject yearly reports. After discussion
in the SIG meeting of 31.10.2023, we have agreed to move it from
a subproject to a "associated repository" of the CAPI
subproject.

A similar case for this can be seen under kubeadm -
e.g. see k/system-validators.