Update community-membership.md
As per https://datatracker.ietf.org/doc/html/rfc2119 , we can replace `may` by `should`  on the new contributors requirements to avoid misenterpretations. 
The spirit of the requirements is to get people that contributes to the project across all areas, if only attend meetings is enough, per example, can be easily gamed.