pull-requests.md: add guidance for large and/or automatic edits
This primarily came out of the discussion around allowing the use of
LLMs (https://github.com/kubernetes/steering/issues/291), but isn't limited to
it because other tools (search/replace, linters) can have the same effect.

The goal is to clarify expected behavior and to give reviewers something that
they can link to when they decide that a PR shouldn't get merged.