Add an optional leads mailing list address
This was inspired by a SIG-Multicluster discussion; there are leads
mailing lists that can be used to contact all the leads of a given
SIG, but they aren't advertised. This adds a field, a template entry,
and the information for SIG-MC.

Since the mailing list is only useful as an email address, the entry
is supposed to be an email address (or slight anti-spam variant
thereof), rather than the web link commonly used for open mailing
lists.

Signed-off-by: Stephen Kitt <skitt@redhat.com>